rename render to detail
