+++ src/features/Conversation/Messages/AssistantGroup/Tool/Detail/RejectedResponse.tsx
